backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "werkzeug>=3.1.5",
]
